Renamed esp folder to esp-libs
