+++ mLRS/Common/common_types.cpp
@@ -9,7 +9,7 @@
 
 #include <string.h>
 #if defined(ESP8266) || defined(ESP32)
-#include "../Common/esp/esp.h"
+#include "../Common/esp-libs/esp.h"
 #else
 #include "../modules/stm32ll-lib/src/stdstm32.h"
 #endif
