+++ mLRS/Common/esp-libs/esp-mcu.h
