+++ platformio.ini
@@ -30,7 +30,7 @@ build_src_filter =
   +<Common/channel_order.cpp>
   +<Common/diversity.cpp>
   +<Common/fhss.cpp>
-  +<Common/esp/esp.c>
+  +<Common/esp-libs/esp.c>
   +<Common/libs/filters.cpp>
 
 [env_common_esp32]
@@ -46,7 +46,7 @@ build_src_filter =
   +<Common/channel_order.cpp>
   +<Common/diversity.cpp>
   +<Common/fhss.cpp>
-  +<Common/esp/esp.c>
+  +<Common/esp-libs/esp.c>
   +<Common/libs/filters.cpp>
 
 [env:rx-dev-900]
@@ -82,7 +82,7 @@ build_src_filter =
 build_flags = 
   -DRX_GENERIC_2400_PA
 
-[env:rx-generic-2400-pa-d]
+[env:rx-generic-2400-pa-td]
 extends = env_common_esp32
 board = esp32dev
 monitor_filters = esp32_exception_decoder, default
